hanging the ssh connection waiting method.
The the current 'wait_for_ssh' method is based on TCP timeout (the default ssh timeout in lago is 10 sec)
when the host is booting, while it doesn't get an IP the timeout is really 10 sec,
but between the time it gets it's IP to the time it opens port 22 the timeout is 0.

So, if one of my vms need more time to open port 22 and start sshd, it might fail after 50 retries.

My fix focus on the maximal 'wait-time' Lago will wait until the host can be reached,
instead of the number of retries that may vary with different OS machines.

Signed-off-by: Ido Rosenzwig <irosenzw@redhat.com>